Updates in preparation for 0.0.4
First pass to update documentation in preparation for 0.0.4.  Mainly focused on the transition from teams -> projects, but also addressed a few other things including rebuilds when new run image is available.  Did not cover the build metadata feature and assumed that users cannot specify a project in their yaml config files.  Maybe we can work off this branch as we continue to add documentation?